Cleaning up the ESZ block
